+++ x/gov/simulation/genesis_test.go
@@ -51,6 +51,7 @@ func TestRandomizedGenState(t *testing.T) {
 		tallyExpeditedThreshold = "0.545000000000000000"
 		tallyVetoThreshold      = "0.327000000000000000"
 		minInitialDepositDec    = "0.880000000000000000"
+		minStakeToVote          = "0stake"
 	)
 
 	assert.Equal(t, "272stake", govGenesis.Params.MinDeposit[0].String())
@@ -62,6 +63,8 @@ func TestRandomizedGenState(t *testing.T) {
 	assert.Equal(t, tallyThreshold, govGenesis.Params.Threshold)
 	assert.Equal(t, tallyExpeditedThreshold, govGenesis.Params.ExpeditedThreshold)
 	assert.Equal(t, tallyVetoThreshold, govGenesis.Params.VetoThreshold)
+	assert.Equal(t, minInitialDepositDec, govGenesis.Params.MinInitialDepositRatio)
+	assert.Equal(t, minStakeToVote, govGenesis.Params.MinStakeToVote.String())
 	assert.Equal(t, uint64(0x28), govGenesis.StartingProposalId)
 	assert.DeepEqual(t, []*v1.Deposit{}, govGenesis.Deposits)
 	assert.DeepEqual(t, []*v1.Vote{}, govGenesis.Votes)